# feat(perf): Improve performance by reducing unnecessary function calls.